Changes to how key events are handled, etc
Added support for keyMod detection when an action key is pressed. So now
things like ctrl+s can be used to save. Changed 'fullscreen' button from
'V' to right-alt+enter. Made the fullscreen change into an Engine
function instead of whatever it was before.

Made the beginnings of a header file for the Caster class, which will be
another module that an Actor can hold.