Finalizes spell creation
This is the first iteration of spell creation. This may radically change
if this system is bad, but it seems to be good so far. I might change
how spells discriminate between enemies, allies, and neutral (as in, not
using TargetType but another field to define how discriminating the
spell is).

I have been neglecting this game because I am in the process of working
on a major project for school (and a corporate sponsor), and I have been
working on a tabletop (pen & paper) rpg.