Changes spell creation a little
This changes how spells are allowed to discriminate between targets,
allowing for more spells. I figure there are around ~24000 possible
spells given the new spell creation system. Not all of them will be
useful (in fact, some of them will definitely be harmful), but at least
there will be a lot of variety.