More spell support, bugfixes
Added a hostility boolean to Actors, indicating whether they are hostile
towards the player or not. Right now, the only Actor not hostile towards
the player is the player itself

Equipment is now unequipped when you drop it. I don't know why I forgot
to implement that before, since I implemented that in Astro.

The spell::cast function now takes the Actor that is casting it as an
argument, which facilitates resolving hostility for spell targeting, as
well as where the spell should originate from. (for targeting a tile)

I will need to modify pickATile so that AI-driven Actors can use it to
target the player (or some other enemy).

attacker.lastTarget should now be handled right again, I think.