Some spell effects implemented
I added a few spell effect implementations. I need to work out the
element system (I plan to have 15 different elements, for the 4
classical elements and their possible combinations).

I also need to implement a decent factory method for creating new actors
based on element or race and "class" or archetype

I plan to also rework the Aura class, as it should be an abstract class
with a bunch of subclasses. The current implementation was a quick fix