Added shields, starting to change Aura class
I am now starting to change the aura class so that each type of aura is
its own unique class. The current state of the Aura class looks horrible
to me overall. I hastily slapped it together in this form for Astro, and
I just kind of left it like that for Caster Edition. Now I am going to
change it (I think for the better).

HP Shields (or overheals) are now implemented. They are essentially a
damage buffer that you can lose before you start taking actual HP
damage. I plan to make the ShieldAura eventually also have an
ElementalSubtype parameter and another float, which will allow for the
shield to give the player elemental resistances (equal to this new float
param) for the duration of the shield.

Also added the Shield SpellEffect functionality