"Fixing" Aura class in progress
I am turning the Aura class into an abstract class and making a subclass
for each possible Aura. I am unsure if that is the best solution, but
that's what I am doing. It adds a bunch more classes but gets rid of the
horrendous switch-case structure of the old Aura class.

I also added a bunch of .clearAndDelete() statements for some of the
TCODLists I am using to make extra-sure that they are deleted when no
longer needed.