Added some more spelleffects, auras
Added absorption and reflection, minus the elemental subtype stuff (I
will get to that part soon, if all goes well. It will take a lot of work
to get everything associated with that implemented).

I also got the Engine function that finds a nearby open tile to work a
little smarter. It takes x and y by reference, and modifies them in
order to return the coordinates of the nearby tile. It seems to be a
smarter solution to me.

I'm not sure if Reflection and Absorption should be Auras, but once
ElementalSubtypes are implemented, the Aura class seems to be a handy
way to store what elements of spells should be absorbed or reflected.

I think I actually found a solution. I plan to store a Destructible
Actor's resistance to each of the 15 elements in a float array of size
15. The indices of the array each correspond to the enum value of a
particular element.
Why not just tack on two more arrays, one with bools (for reflection)
and one with floats (absorption)? It seems kind of hack-y to use enums
for their integer values, but I've seen source where people use and enum
that's called like NB_OF_ELEMENTS to use as a count of all possible
enumerated values, so I'm not the only one who has done it before. That
doesn't make it better, but I digress...