Added a ShuffleBag class
The ShuffleBag class is a class that is meant to help implement filtered
randomness. You essentially fill the shufflebag with specific values
that you want to be randomly selected from, and it retrieves them for
you.

Let's say you have 12 possible places you can spawn an enemy. But you
don't want them to be able to spawn any place more than 3 times. You
fill the shufflebag with 1-12 three times, for a total of 36 items, and
then you just retrieve an item from the shufflebag until its empty. This
way, the enemies will spawn in all 12 spawn points "randomly," at a max
of 3 times in one period. If you want to keep spawning enemies, simply
fill up the shufflebag again and keep going.