Few more spelleffects implemented
I also made some changes to the map class. I am thinking I will use the
map class more like the way its intended to be used, as an area equal to
the visible area of the actual map. If I restrict it in size this way,
computing the FOV becomes much more efficient, as it doesn't have to
examine the full size of the actual map.