Got rid of item creation, added controlled teleport
Creating items at will seemed too strong of a spell to me. Perhaps I
might add it as a consumable scroll that creates random artifacts. I
might add it later as a spell for the player that creates mostly junk
but might create an artifact at a small chance, and maybe occasionally
consumables.

Controlled teleport simply lets the caster choose a destination for each
target of the spell. This may be tedious if it is a level-wide spell,
but maybe the player won't want to cast that too often because of it.
I'd rather let them choose a target for every teleport target than just
flood a single tile with the entire contents of the level.

All SpellEffects for CreatureSpells now have an entry in the case-switch
structure, but some of them require more features to be added to the
game to have a meaningful implementation. I will soon be adding a speed
system, more buffs and debuffs, revamping Weapons to handle multiple
damage types and enhancements (and possibly changing the Attack->attack
function to handle these changes), and adding a natural attack slot to
the Container class so that all Actors that can attack have a default
melee weapon.