Beginnings of the caster class being added
Good amount of support for the Caster class is being added to the
source. Much like Pickable, Attacker, Container, etc..., Caster is a
module that can be tacked onto an Actor in order to give it the ability
to cast spells.

The player can press ctrl+c to enter their Cantus, which is necessary to
cast spells. More elaboration will come as spellcasting comes closer to
fruition.

I am working to get to the point where the player can cast even a single
spell, in order to being testing those effects which are actually in
place. After which point, I will begin implementing the features
necessary to support the other possible spellEffects.