Much of the spell instability is gone
There are a good number of commented-out std::out calls and a few
include <iostream> statements that I will take out soon, but I want to
make sure that everything is functioning right regarding spells.

Additionally, the game now uses a heck of a lot less CPU! Turns out, I
was previously using the libtcod call for the non-blocking SDL input
call. So SDL was running on a tight loop constantly polling for input,
instead of giving up the CPU and only taking it back when input is
given. Since I changed that, it typically uses 0-0.1% of my CPU, which
is much better than 8-16%. I'm very glad I was able to quash the
resource hogging, as it made no sense to me.