Can now navigate the spellbook
The spellbook can now be navigated page-by-page if the caster has more
than 26 spells (we use lowercase a-z to denote possible choices from the
spellbook, just as with the inventory).

Using left arrow or keypad-4 will go back a page, and right arrow or
keypad-6 will go forward a page.

Spells costs are now also flat numbers, not mixed decimal and integer