Closing the spellbook is a little easier
I figured hitting escape to close the spellbook might be too much of a
hand movement for some players. As such, now you can also close the
spellbook with either the down arrow or numpad-2/keypad-2. The escape
key also works, so that's three methods to close the spellbook, which
should be enough to minimize inconvenience for most people.

Alternately, I could make it so that any key that isn't a-z will close
the spellbook, so if people (hah) would prefer that, let me know.