New Scheduler idea
Now my idea for the scheduler is as follows: every actor that can
potentially update itself must have an Ai module and a new module called
Speedy, which simply holds its speed and maxSpeed. Speed is a value for
the number of turns the actor should take per initiative pass.
Essentially, if an Actor has a speed of 5, it will take 5 turns per
every iteration of Engine::update(). These 5 turns will be mixed in with
every other Actor's turns, although some of their earlier turns may
happen one-after-the-other since the scheduler uses a max binary heap to
store the actors based on their speeds.

This commit adds the stuff necessary for the scheduler, but does not add
the Speedy class, modify Actor to store an instance of Speedy, or modify
Engine::update() to use the scheduler.

I might or might not add those tonight, as I have an exam tomorrow
morning, but they should definitely be added by the end of tomorrow
night.

PS: I really dont know what messed up the other scheduler. If anyone
(hah) has an idea as to what went wrong, please let me know.