Clean up the scheduler a little bit
This is a bit of a rushed addition, as I want to get back to studying,
but I realized that I forgot to account for Actors that die in the
course of the game. I figure they should be removed from the scheduler,
as they can't update while dead.

Of course, actors that are resurrected should be pushed back into the
scheduler. Any Actors that are created after the level begins (spell
effects, etc) should also be pushed into the scheduler