Final Commit for a long while, and after a long while
Big necro commit to immortalize the partial rework that I have visited
briefly over the past few years. I got really busy with real life when I
was last committing actively, and eventually I dropped this project
entirely. I spent a few hours strung across three years to try and begin
a rewrite of the spellcasting system to remove some of the switch-case
statements (or at least move them just to factory classes), and I think
i was on the right track.

Just wanted to immortalize the rework and sort of officially sign off of
this project.

Something new is in the works, however, so don't worry.